requirements.txt
Flask
numpy==1.19.5
regex==2020.11.13
keras
tensorflow-cpu
gunicorn

